Modify vu32 type to fit normal use of uint32 in 32-bit MCU

+++ Library/Device/Nuvoton/NUC1263/Include/NUC1263.h
@@ -320,7 +320,7 @@ extern void SystemInit(void);
 
 //=============================================================================
 typedef volatile unsigned char  vu8;
-typedef volatile unsigned long  vu32;
+typedef volatile unsigned int   vu32;
 typedef volatile unsigned short vu16;
 #define M8(adr)  (*((vu8  *) (adr)))
 #define M16(adr) (*((vu16 *) (adr)))
